Implemented Delete Btn

+++ static/app.js
@@ -141,7 +141,7 @@ async function fetchHabits() {
         
         habitItem.addEventListener("click", () => {
           // if (!habitItem.classList.contains("done")) {
-          trackHabit(habit.habit_id);
+          // trackHabit(habit.habit_id);
           // }
         });
 
@@ -162,10 +162,12 @@ async function fetchHabits() {
           uncheckButton.style.display = "none";
           deleteButton.style.display = "flex";
           editButton.style.display = "flex";
+
+          deleteButton.addEventListener("click", () => deleteHabit(habit));
         }
       });
     } else {
-      // Handle fetch habits error
+      // TODO Handle fetch habits error
     }
   } catch (error) {
     console.error("Error fetching habits:", error);
@@ -173,19 +175,22 @@ async function fetchHabits() {
   }
 }
 
-async function deleteHabit(habitId) {
+async function deleteHabit(habit) {
   try {
-    const response = await fetch(`/habit/${habitId}`, {
-      method: "DELETE",
-    });
+    const confirmed = confirm(`Are you sure you want to delete ${habit.name}?`);
+    if (confirmed) {
+      const response = await fetch(`/habit/${habit.habit_id}`, {
+        method: "DELETE",
+      });
 
-    if (response.ok) {
-      // Refresh the habit list after deleting
-      fetchHabits();
-    } else {
-      // Handle delete habit error
-    }
-
+      if (response.ok) {
+        // Refresh the habit list after deleting
+        fetchHabits();
+      } else {
+        const errorData = await response.json(); // Assuming the server sends a JSON error response
+        console.error("Delete failed:", errorData.message);
+      }
+    } 
     // exitDeleteMode();
   } catch (error) {
     console.error("Error deleting habit:", error);
